Add Gradio support in one-by-one executor too.
