+++ lynxkite-core/pyproject.toml
@@ -13,8 +13,8 @@ classifiers = ["License :: OSI Approved :: GNU Affero General Public License v3 
 Homepage = "https://github.com/lynxkite/lynxkite-2000/"
 
 [tool.deptry.per_rule_ignores]
-DEP001 = ["matplotlib", "griffe", "pycrdt"]
-DEP003 = ["matplotlib", "griffe", "pycrdt"]
+DEP001 = ["matplotlib", "griffe", "pycrdt", "gradio", "starlette"]
+DEP003 = ["matplotlib", "griffe", "pycrdt", "gradio", "starlette"]
 
 [build-system]
 requires = ["uv_build>=0.8.0,<0.9.0"]
