+++ lynxkite-graph-analytics/pyproject.toml
@@ -50,9 +50,7 @@ name = "nvidia"
 url = "https://pypi.nvidia.com"
 
 [tool.deptry.per_rule_ignores]
-DEP001 = ["gradio", "starlette"]
 DEP002 = ["numba", "pyarrow", "nx-cugraph-cu12", "pylibcugraph-cu12"]
-DEP003 = ["gradio", "starlette"]
 
 [tool.deptry.package_module_name_map]
 grand-cypher = "grandcypher"
